% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Application UI}
\usage{
app_ui()
}
\value{
A UI definition for a Shiny app
}
\description{
Constructs the user interface using \code{bslib} components.
}
\note{
last edited: 2025-06-14-23.23.36
}
