% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_input_server}
\alias{mod_input_server}
\title{Variable Input Server Module}
\usage{
mod_input_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A list with the reactive data from \code{load_tt_data()} and the dataset title.
}
\description{
Variable Input Server Module
}
