% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_input_ui}
\alias{mod_input_ui}
\title{Variable Input UI Module}
\usage{
mod_input_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI elements for dataset title selection
}
\description{
Variable Input UI Module
}
