% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_list.R
\name{mod_list_server}
\alias{mod_list_server}
\title{Table Server Module (listviewerlite)}
\usage{
mod_list_server(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{A reactive expression returning the dataset list}
}
\description{
Table Server Module (listviewerlite)
}
