% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_list.R
\name{mod_list_ui}
\alias{mod_list_ui}
\title{List UI Module (listviewerlite)}
\usage{
mod_list_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI element for list output using listviewerlite
}
\description{
List UI Module (listviewerlite)
}
