% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Inspect Plot Module Server}
\usage{
mod_plot_server(id, ttd)
}
\arguments{
\item{id}{Module id}

\item{ttd}{Reactive or a static named list of data.frames}
}
\description{
Renders one plot per (dataset × plot-type)
}
