% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_ui}
\alias{mod_plot_ui}
\title{Inspect Plot Module UI}
\usage{
mod_plot_ui(id)
}
\arguments{
\item{id}{Module id}
}
\value{
Shiny UI
}
\description{
Let the user pick plots and see one panel per dataset × plot
}
