% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_desc.R
\name{mod_report_desc_server}
\alias{mod_report_desc_server}
\title{Report Description Format Server Module}
\usage{
mod_report_desc_server(id, format)
}
\arguments{
\item{id}{Module ID}

\item{format}{A reactive expression returning the report format}
}
\description{
Report Description Format Server Module
}
