% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_desc.R
\name{mod_report_desc_ui}
\alias{mod_report_desc_ui}
\title{Report Description UI Module with Format Selection}
\usage{
mod_report_desc_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI elements for report format descriptions
}
\description{
Report Description UI Module with Format Selection
}
