% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_input.R
\name{mod_report_input_server}
\alias{mod_report_input_server}
\title{Report Input Server Module}
\usage{
mod_report_input_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A reactive list with the selected format string
}
\description{
Report Input Server Module
}
