% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_input.R
\name{mod_report_input_ui}
\alias{mod_report_input_ui}
\title{Report Input UI Module with Format Selection}
\usage{
mod_report_input_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI elements for report format input
}
\description{
Report Input UI Module with Format Selection
}
