% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report Server Module with Enhanced Quarto Error Handling}
\usage{
mod_report_server(id, data, selected_plot_type, dataset_title)
}
\arguments{
\item{id}{Module ID}

\item{data}{A reactive expression returning the dataset list}

\item{selected_plot_type}{A reactive expression returning the selected plot type}

\item{dataset_title}{A reactive expression returning the dataset title}
}
\description{
Report Server Module with Enhanced Quarto Error Handling
}
