% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Report UI Module with Format Selection}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI elements for report generation with format options
}
\description{
Report UI Module with Format Selection
}
