% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{Table Server Module (reactable)}
\usage{
mod_table_server(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{A reactive expression returning the dataset list}
}
\description{
Table Server Module (reactable)
}
