% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_ui}
\alias{mod_table_ui}
\title{Table UI Module (reactable)}
\usage{
mod_table_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI element for data table output using reactable
}
\description{
Table UI Module (reactable)
}
