% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_viz.R
\name{mod_viz_server}
\alias{mod_viz_server}
\title{Visualization Server Module (inspectdf-compatible) - Enhanced}
\usage{
mod_viz_server(id, data)
}
\arguments{
\item{id}{Module ID}

\item{data}{A reactive expression returning the dataset list}
}
\value{
A reactive value containing the current plot type
}
\description{
Visualization Server Module (inspectdf-compatible) - Enhanced
}
