% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_viz.R
\name{mod_viz_ui}
\alias{mod_viz_ui}
\title{Visualization UI Module (inspectdf-compatible)}
\usage{
mod_viz_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\value{
UI output for one or more inspectdf plots
}
\description{
Visualization UI Module (inspectdf-compatible)
}
