% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_available.R
\name{quarto_available}
\alias{quarto_available}
\title{Check Quarto Availability}
\usage{
quarto_available()
}
\value{
Logical \code{TRUE} if Quarto is available, \code{FALSE} otherwise.
}
\description{
Checks whether Quarto is available via the R package, system path, or
common installation locations.
}
\keyword{internal}
