% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R, R/render_rmarkdown_report.R
\name{render_rmarkdown_report}
\alias{render_rmarkdown_report}
\title{Render R Markdown Report}
\usage{
render_rmarkdown_report(template_path, output_file, params)

render_rmarkdown_report(template_path, output_file, params)
}
\arguments{
\item{template_path}{Path to the \code{.Rmd} template file.}

\item{output_file}{Output file path.}

\item{params}{Named list of report parameters.}
}
\value{
Called for its side-effect of writing \code{output_file}.
}
\description{
Renders an R Markdown (\code{.Rmd}) template to the specified output file
using \code{rmarkdown::render()}.
}
\keyword{internal}
