% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_logger.R, R/testthat.R
\name{test_logger}
\alias{test_logger}
\title{Test Logger (Test Utility)}
\usage{
test_logger(start = NULL, end = NULL, msg)

test_logger(start = NULL, end = NULL, msg)
}
\arguments{
\item{start}{test start message}

\item{end}{test end message}

\item{msg}{test message}
}
\value{
Called for its side-effect of writing to the log output.

message to test output
}
\description{
Emits structured log messages to the test output, suitable for bracketing
test blocks with start/end markers.
}
\keyword{internal}
