% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R, R/tstmp.R
\name{tstmp}
\alias{tstmp}
\title{Timestamp}
\usage{
tstmp()

tstmp()
}
\value{
character string of date/time.

A character string of the current date and time.
}
\description{
Returns the current system time formatted as \code{"YYYY-MM-DD-HH.MM.SS"}.
}
\examples{
tstmp()
cat(paste("Last updated:", tstmp()))
tstmp()
cat(paste("Last updated:", tstmp()))

}
