% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{convert_qmd_to_rmd}
\alias{convert_qmd_to_rmd}
\title{Convert Quarto syntax to R Markdown}
\usage{
convert_qmd_to_rmd(qmd_lines)
}
\description{
Convert Quarto syntax to R Markdown
}
\keyword{internal}
