% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{create_fallback_qmd_template}
\alias{create_fallback_qmd_template}
\title{Create Fallback Quarto Template}
\usage{
create_fallback_qmd_template()
}
\description{
Create Fallback Quarto Template
}
\keyword{internal}
