% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{create_fallback_rmd_template}
\alias{create_fallback_rmd_template}
\title{Create Fallback R Markdown Template}
\usage{
create_fallback_rmd_template()
}
\description{
Create Fallback R Markdown Template
}
\keyword{internal}
