% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{create_fallback_template}
\alias{create_fallback_template}
\title{Create Fallback Template}
\usage{
create_fallback_template(format = c("rmarkdown", "quarto"))
}
\arguments{
\item{format}{Report format}
}
\value{
Path to created template
}
\description{
Create Fallback Template
}
\keyword{internal}
