% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tt_data.R
\name{extract_file_names}
\alias{extract_file_names}
\title{Extract file names from TidyTuesday data object}
\usage{
extract_file_names(tt_data, preserve_original_names = FALSE)
}
\arguments{
\item{tt_data}{list of TidyTuesday data from \code{\link[=get_tt_data]{get_tt_data()}}}

\item{preserve_original_names}{Logical, fallback to original names}
}
\value{
Character vector of file names with .gz extensions removed
}
\description{
Extract file names from TidyTuesday data object
}
\examples{
ttd <- get_tt_data("Moore’s Law")
extract_file_names(ttd)

}
