% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{get_template_path}
\alias{get_template_path}
\title{Get Template Path for Format}
\usage{
get_template_path(format)
}
\arguments{
\item{format}{Report format ("rmarkdown" or "quarto")}
}
\value{
Path to template file
}
\description{
Get Template Path for Format
}
\keyword{internal}
