% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_download.R
\name{mod_report_download_server}
\alias{mod_report_download_server}
\title{Report Download Server Module with Enhanced Quarto Error Handling}
\usage{
mod_report_download_server(id, format, data, selected_plot_type, dataset_title)
}
\arguments{
\item{id}{Module ID}

\item{format}{A reactive expression returning the report format}

\item{data}{A reactive expression returning the dataset list}

\item{selected_plot_type}{A reactive expression returning the selected plot type}

\item{dataset_title}{A reactive expression returning the dataset title}
}
\description{
Report Download Server Module with Enhanced Quarto Error Handling
}
