% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{quarto_available}
\alias{quarto_available}
\title{Quarto availability check}
\usage{
quarto_available()
}
\value{
Logical indicating if Quarto is available
}
\description{
Quarto availability check
}
\keyword{internal}
