% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_quarto_alternative}
\alias{render_quarto_alternative}
\title{Alternative Quarto Rendering using rmarkdown}
\usage{
render_quarto_alternative(template_path, output_file, params)
}
\description{
Alternative Quarto Rendering using rmarkdown
}
\keyword{internal}
