% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_quarto_system_call}
\alias{render_quarto_system_call}
\title{System call method for Quarto rendering}
\usage{
render_quarto_system_call(template_path, output_file, params)
}
\arguments{
\item{template_path}{path to template}

\item{output_file}{report output}
}
\description{
System call method for Quarto rendering
}
\keyword{internal}
