% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_rmarkdown_report}
\alias{render_rmarkdown_report}
\title{Render R Markdown Report}
\usage{
render_rmarkdown_report(template_path, output_file, params)
}
\arguments{
\item{template_path}{Path to Rmd template}

\item{output_file}{Output file path}

\item{params}{Report parameters}
}
\description{
Render R Markdown Report
}
\keyword{internal}
