% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_plot.R
\name{ttd_length}
\alias{ttd_length}
\title{Determine length of TidyTuesday Data}
\usage{
ttd_length(ttd)
}
\arguments{
\item{ttd}{A list of data frames from \code{\link[=load_tt_data]{load_tt_data()}}}
}
\value{
integer length of \code{ttd}
}
\description{
Determine length of TidyTuesday Data
}
\examples{
ttd <- load_tt_data("Moore’s Law")
ttd_length(ttd)

}
