% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{validate_and_clean_params}
\alias{validate_and_clean_params}
\title{Validate and Clean Report Parameters}
\usage{
validate_and_clean_params(params)
}
\arguments{
\item{params}{List of parameters}
}
\value{
Cleaned and validated parameter list
}
\description{
Validate and Clean Report Parameters
}
\keyword{internal}
