% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{validate_quarto_params}
\alias{validate_quarto_params}
\title{Validate Quarto Parameters}
\usage{
validate_quarto_params(params)
}
\arguments{
\item{params}{Parameter list}
}
\value{
Validated parameter list
}
\description{
Validate Quarto Parameters
}
\keyword{internal}
